+++ python/sdss_brain/helpers/io.py
@@ -1,26 +1,26 @@
 # !/usr/bin/env python
 # -*- coding: utf-8 -*-
-# 
-# Filename: helpers.py
-# Project: sdss_brain
+#
+# Filename: io.py
+# Project: helpers
 # Author: Brian Cherinka
-# Created: Monday, 16th March 2020 1:00:14 pm
+# Created: Wednesday, 7th October 2020 10:54:12 am
 # License: BSD 3-clause "New" or "Revised" License
 # Copyright (c) 2020 Brian Cherinka
-# Last Modified: Monday, 16th March 2020 5:29:39 pm
+# Last Modified: Wednesday, 7th October 2020 10:54:12 am
 # Modified By: Brian Cherinka
 
 
-from __future__ import absolute_import, division, print_function
-
+from __future__ import print_function, division, absolute_import
 import pathlib
+from typing import Union
 from astropy.io import fits
 from sdss_brain import log
 from sdss_brain.config import config
 from sdss_brain.exceptions import BrainError
 
 
-def get_mapped_version(name, release=None, key=None):
+def get_mapped_version(name: str, release: str = None, key: str = None) -> Union[dict, str]:
     ''' Get a version id mapped to a release number
 
     For a given named category, looks up the "mapped_versions" attribute from
@@ -29,22 +29,29 @@ def get_mapped_version(name, release=None, key=None):
     and dapver='2.2.1'. This can be useful when needing to specify certain versions
     when defining paths to files.
 
-    Parameters:
-        name (str):
+    Parameters
+    ----------
+        name : str
             The name of the set of versions to access
-        release (str):
+        release : str
             The SDSS release.  Default is config.release.
-        key (str):
+        key : str
             Optional name of dictionary key to access specific value
 
-    Example:
+    Returns
+    -------
+        version : dict|str
+            A version number corresponding to a given release
+
+    Example
+    -------
         >>> # access the MaNGA versions for release DR16
         >>> get_mapped_version('manga', release='DR16')
-        >>> {'drpver': 'v2_4_3', 'dapver': '2.2.1'}
-        >>> 
+            {'drpver': 'v2_4_3', 'dapver': '2.2.1'}
+
         >>> # access specific key
         >>> get_mapped_version('manga', release='DR16', key='drpver')
-        >>> 'v2_4_3'
+            'v2_4_3'
     '''
 
     # get the mapped_versions attribute from the configuration
@@ -74,17 +81,20 @@ def get_mapped_version(name, release=None, key=None):
     return version
 
 
-def load_fits_file(filename):
+def load_fits_file(filename: str) -> fits.HDUList:
     ''' Load a FITS file
 
     Opens and loads a FITS file with astropy.io.fits.
 
-    Parameters:
-        filename (str):
+    Parameters
+    ----------
+        filename : str
             A FITS filen to open
 
-    Returns:
-        an Astropy HDUList
+    Returns
+    -------
+        hdulist : `~astropy.io.fits.HDUList`
+            an Astropy HDUList
     '''
 
     path = pathlib.Path(filename)
@@ -99,4 +109,3 @@ def load_fits_file(filename):
         raise BrainError(f'Failed to open FITS files {filename}: {err}')
     else:
         return hdulist
-